requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
resend>=2.0.0
